Added Back End Files
These are all the header files I've made for the game so far. everything is perfect with the exception of:

BackEnd/StraightSpot.h,
which occasionally struggles to tell when you have certain straights,

BackEnd/BotPlayer.h,
which sometimes struggles when going for a straight.

Other than those small things i may also improve the bots decision making to make it score higher.